+++ src/hooks/mutations/useEditMemberInfoMutation.ts
@@ -1,7 +1,6 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation } from "@tanstack/react-query";
 import { toast } from "react-toastify";
 import { allMemberApi } from "@/apis/allMemberApi";
-import { QueryKey } from "@/constants/queryKey";
 
 type EditMemberInfoMutationArgumentType = {
   memberId: number;
@@ -14,24 +13,12 @@ type EditMemberInfoMutationArgumentType = {
     discordUsername: string | null;
     nickname: string | null;
   };
-  onSuccess: () => void;
 };
 
 export default function useEditMemberInfoMutation() {
-  const queryClient = useQueryClient();
-
   return useMutation({
     mutationFn: ({ memberId, body }: EditMemberInfoMutationArgumentType) =>
       allMemberApi.editMemberInfo(memberId, body),
-    onSuccess: async (_, variables) => {
-      Promise.all([
-        queryClient.invalidateQueries({
-          queryKey: [QueryKey.allMemberList],
-        }),
-      ]);
-      toast.success("수정 완료되었습니다.");
-      variables.onSuccess();
-    },
     onError: (error: any) => {
       toast.error(error.response.data.errorMessage);
     },
