+++ src/hooks/mutations/useEditRecruitmentRoundMutation.ts
@@ -1,11 +1,8 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation } from "@tanstack/react-query";
 import { toast } from "react-toastify";
 import { recruitmentApi } from "@/apis/recruitmentApi";
-import { QueryKey } from "@/constants/queryKey";
 
 export default function useEditRecruitmentRoundMutation() {
-  const queryClient = useQueryClient();
-
   return useMutation({
     mutationFn: ({
       recruitmentRoundId,
@@ -19,12 +16,6 @@ export default function useEditRecruitmentRoundMutation() {
         roundType: "FIRST" | "SECOND";
       };
     }) => recruitmentApi.editRecruitmentRound(recruitmentRoundId, body),
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: [QueryKey.recruitmentRound],
-      });
-      toast.success("모집 회차 정보가 수정되었습니다.");
-    },
     onError: (error: any) => {
       toast.error(error.response.data.errorMessage);
     },
